receiver/requirements.txt

confluent_kafka
kafka-python
mysql-connector-python
